refactor: rename configutations folder to configurations
