+++ src/configurations/remotePlugin.test.ts
