+++ src/configurations/remotePlugin.ts
