+++ src/index.ts
@@ -3,8 +3,8 @@ import path from 'path'
 import {mergeDeepRight} from 'rambda'
 import {createUnplugin} from 'unplugin'
 
-import {retrieveHostConfig} from './configutations/hostPlugin'
-import {retrieveRemoteConfig} from './configutations/remotePlugin'
+import {retrieveHostConfig} from './configurations/hostPlugin'
+import {retrieveRemoteConfig} from './configurations/remotePlugin'
 import {HostOptions} from './interfaces/HostOptions'
 import {RemoteOptions} from './interfaces/RemoteOptions'
 import {createTypesArchive, downloadTypesArchive} from './lib/archiveHandler'
